# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UsersController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me": {
      "patch": {
        "operationId": "UsersController_updateMe_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOwnProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/master/module-systems": {
      "get": {
        "operationId": "ModuleSystemsController_listSystems_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ModuleSystems"
        ]
      },
      "post": {
        "operationId": "ModuleSystemsController_createSystem_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertModuleSystemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ModuleSystems"
        ]
      }
    },
...
    "/api/v1/master/module-systems/{systemId}": {
      "get": {
        "operationId": "ModuleSystemsController_getSystem_v1",
        "parameters": [
          {
            "name": "systemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ModuleSystems"
        ]
      },
      "patch": {
        "operationId": "ModuleSystemsController_updateSystem_v1",
        "parameters": [
          {
            "name": "systemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertModuleSystemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ModuleSystems"
        ]
      }
    },
    "/api/v1/master/module-systems/{systemId}/status": {
      "patch": {
        "operationId": "ModuleSystemsController_updateStatus_v1",
        "parameters": [
          {
            "name": "systemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateModuleSystemStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ModuleSystems"
        ]
      }
    },
    "/api/v1/master/module-systems/{systemId}/modules": {
      "put": {
        "operationId": "ModuleSystemsController_saveModules_v1",
        "parameters": [
          {
            "name": "systemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveModuleSystemModulesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ModuleSystems"
        ]
      }
    },
    "/api/v1/master/module-systems/{systemId}/sidebar": {
      "get": {
        "operationId": "ModuleSystemsController_getSidebar_v1",
        "parameters": [
          {
            "name": "systemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ModuleSystems"
        ]
      },
      "put": {
        "operationId": "ModuleSystemsController_saveSidebar_v1",
        "parameters": [
          {
            "name": "systemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveModuleSystemSidebarDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ModuleSystems"
        ]
      }
    },
...
    "/api/v1/workspace/users": {
      "get": {
        "operationId": "WorkspaceUsersController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceUserResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Workspace Users"
        ]
      },
      "post": {
        "operationId": "WorkspaceUsersController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceUserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workspace Users"
        ]
      }
    },
    "/api/v1/workspace/users/{userId}/resend-invitation": {
      "post": {
        "operationId": "WorkspaceUsersController_resendInvitation_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceUserMessageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workspace Users"
        ]
      }
    },
    "/api/v1/workspace/users/{userId}/invitation": {
      "delete": {
        "operationId": "WorkspaceUsersController_cancelInvitation_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceUserCancelInvitationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workspace Users"
        ]
      }
    },
    "/api/v1/workspace/users/{userId}": {
      "patch": {
        "operationId": "WorkspaceUsersController_update_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceUserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workspace Users"
        ]
      }
    },
...
    "/api/v1/table-preferences/{moduleKey}": {
      "get": {
        "operationId": "TablePreferencesController_findOne_v1",
        "parameters": [
          {
            "name": "moduleKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Table preference retrieved."
          }
        },
        "tags": [
          "Table Preferences"
        ]
      },
      "put": {
        "operationId": "TablePreferencesController_save_v1",
        "parameters": [
          {
            "name": "moduleKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveTablePreferenceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Table preference saved."
          }
        },
        "tags": [
          "Table Preferences"
        ]
      }
    },
...
  },
  "info": {
    "title": "GR8Books Neo API",
    "description": "Versioned REST API for GR8Books Neo frontend clients.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "contactNumber": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "systemRole": {
            "enum": [
              "SUPER_ADMIN",
              "STANDARD"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "UpdateOwnProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "maxLength": 120
          },
          "contactNumber": {
            "type": "string",
            "maxLength": 40
          }
        }
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "contactNumber": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "systemRole": {
            "enum": [
              "SUPER_ADMIN",
              "STANDARD"
            ],
            "type": "string"
          }
        }
      },
...
      "UpsertModuleSystemDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64
          },
          "name": {
            "type": "string",
            "maxLength": 160
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 1000
          },
          "sortOrder": {
            "type": "number",
            "minimum": 0
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
      "UpdateModuleSystemStatusDto": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "isActive"
        ]
      },
      "SaveModuleSystemModulesDto": {
        "type": "object",
        "properties": {
          "moduleCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "moduleCodes"
        ]
      },
      "ModuleSystemSidebarItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "key": {
            "type": "string",
            "maxLength": 120
          },
          "label": {
            "type": "string",
            "maxLength": 160
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 500
          },
          "itemType": {
            "type": "string",
            "enum": [
              "SECTION",
              "CONTAINER",
              "LINK"
            ]
          },
          "moduleId": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "moduleCode": {
            "type": "string",
            "nullable": true,
            "maxLength": 120
          },
          "iconName": {
            "type": "string",
            "nullable": true,
            "maxLength": 120
          },
          "sortOrder": {
            "type": "number",
            "minimum": 0
          },
          "isVisible": {
            "type": "boolean"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleSystemSidebarItemDto"
            }
          }
        },
        "required": [
          "key",
          "label",
          "itemType",
          "children"
        ]
      },
      "SaveModuleSystemSidebarDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleSystemSidebarItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
      "WorkspaceUserAssignedUnitResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "companyId": {
            "type": "number"
          },
          "type": {
            "enum": [
              "HEAD_OFFICE",
              "BRANCH",
              "SATELLITE"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "companyId",
          "type",
          "name",
          "displayName",
          "isActive"
        ]
      },
      "WorkspaceUserCompanyAssignmentResponseDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "number"
          },
          "unitIds": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "units": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceUserAssignedUnitResponseDto"
            }
          }
        },
        "required": [
          "companyId",
          "unitIds",
          "units"
        ]
      },
      "WorkspaceUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "contactNumber": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "lastLogin": {
            "type": "string",
            "nullable": true
          },
          "profileImageUrl": {
            "type": "string",
            "nullable": true
          },
          "companyAssignments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceUserCompanyAssignmentResponseDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "contactNumber",
          "status",
          "lastLogin",
          "profileImageUrl",
          "companyAssignments",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "WorkspaceUserAssignmentDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "number",
            "minimum": 1
          },
          "unitIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "number",
              "minimum": 1
            }
          }
        },
        "required": [
          "companyId",
          "unitIds"
        ]
      },
      "CreateWorkspaceUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "contactNumber": {
            "type": "string"
          },
          "companyAssignments": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceUserAssignmentDto"
            }
          }
        },
        "required": [
          "name",
          "email",
          "companyAssignments"
        ]
      },
      "WorkspaceUserMessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "WorkspaceUserCancelInvitationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "message"
        ]
      },
      "UpdateWorkspaceUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "contactNumber": {
            "type": "string"
          },
          "companyAssignments": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceUserAssignmentDto"
            }
          }
        },
        "required": [
          "name",
          "email",
          "companyAssignments"
        ]
      },
...
      "SaveTablePreferenceDto": {
        "type": "object",
        "properties": {
          "configuration": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "configuration"
        ]
      },
...
    }
  }
}